chore: move evm functions to utils folder
