+++ apps/tangle-dapp/constants/index.ts
@@ -1,1 +1,4 @@
-export * from './evm';
+export const PAYMENT_DESTINATION_OPTIONS = [
+  'Staked (increase the amount at stake)',
+  'Stash (do not increase the amount at stake)',
+];
