+++ apps/tangle-dapp/utils/evm/bond.ts
@@ -1,23 +1,13 @@
-import { u8aToHex } from '@polkadot/util';
-import { decodeAddress } from '@polkadot/util-crypto';
 import { ensureHex } from '@webb-tools/dapp-config';
-import { chainsConfig } from '@webb-tools/dapp-config/chains/chain-config';
 import { AddressType } from '@webb-tools/dapp-config/types';
-import { PresetTypedChainId } from '@webb-tools/dapp-types/ChainId';
-import {
-  createPublicClient,
-  createWalletClient,
-  custom,
-  defineChain,
-  http,
-  parseEther,
-} from 'viem';
+import { parseEther } from 'viem';
 
-import { PaymentDestination } from '../types';
 import {
   StakingInterfacePrecompileABI,
   StakingInterfacePrecompileAddress,
-} from './contract';
+} from '../../constants/contract';
+import { PaymentDestination } from '../../types';
+import { createEvmWalletClient, evmPublicClient } from './client';
 
 const PAYEE_STAKED =
   '0x0000000000000000000000000000000000000000000000000000000000000000';
@@ -25,28 +15,6 @@ const PAYEE_STASH =
   '0x0000000000000000000000000000000000000000000000000000000000000001';
 const PAYEE_CONTROLLER =
   '0x0000000000000000000000000000000000000000000000000000000000000002';
-
-export const PAYMENT_DESTINATION_OPTIONS = [
-  'Staked (increase the amount at stake)',
-  'Stash (do not increase the amount at stake)',
-];
-
-const tangleTestnetConfig = chainsConfig[PresetTypedChainId.TangleTestnetEVM];
-delete tangleTestnetConfig.contracts;
-
-const tangleTestnet = defineChain(tangleTestnetConfig);
-
-export const evmPublicClient = createPublicClient({
-  chain: tangleTestnet,
-  transport: http(),
-});
-
-const createEvmWalletClient = (nominatorAddress: string) =>
-  createWalletClient({
-    chain: tangleTestnet,
-    account: ensureHex(nominatorAddress),
-    transport: custom(window.ethereum),
-  });
 
 export const bondTokens = async (
   nominatorAddress: string,
@@ -87,55 +55,6 @@ export const bondExtraTokens = async (
     abi: StakingInterfacePrecompileABI,
     functionName: 'bondExtra',
     args: [value],
-    account: ensureHex(nominatorAddress),
-  });
-
-  const evmWalletClient = createEvmWalletClient(nominatorAddress);
-
-  const txHash = await evmWalletClient.writeContract(request);
-
-  return txHash;
-};
-
-export const nominateValidators = async (
-  nominatorAddress: string,
-  validatorAddresses: string[]
-): Promise<AddressType> => {
-  const targets = validatorAddresses.map((address) => {
-    return u8aToHex(decodeAddress(address));
-  });
-
-  const { request } = await evmPublicClient.simulateContract({
-    address: StakingInterfacePrecompileAddress,
-    abi: StakingInterfacePrecompileABI,
-    functionName: 'nominate',
-    args: [targets],
-    account: ensureHex(nominatorAddress),
-  });
-
-  const evmWalletClient = createEvmWalletClient(nominatorAddress);
-
-  const txHash = await evmWalletClient.writeContract(request);
-
-  return txHash;
-};
-
-export const updatePaymentDestination = async (
-  nominatorAddress: string,
-  paymentDestination: string
-): Promise<AddressType> => {
-  const payee =
-    paymentDestination === PaymentDestination.Staked
-      ? PAYEE_STAKED
-      : paymentDestination === PaymentDestination.Stash
-      ? PAYEE_STASH
-      : PAYEE_CONTROLLER;
-
-  const { request } = await evmPublicClient.simulateContract({
-    address: StakingInterfacePrecompileAddress,
-    abi: StakingInterfacePrecompileABI,
-    functionName: 'setPayee',
-    args: [payee],
     account: ensureHex(nominatorAddress),
   });
 
@@ -188,14 +107,15 @@ export const rebondTokens = async (
   return txHash;
 };
 
-export const stopNomination = async (
-  nominatorAddress: string
+export const withdrawUnbondedTokens = async (
+  nominatorAddress: string,
+  slashingSpans: number
 ): Promise<AddressType> => {
   const { request } = await evmPublicClient.simulateContract({
     address: StakingInterfacePrecompileAddress,
     abi: StakingInterfacePrecompileABI,
-    functionName: 'chill',
-    args: [],
+    functionName: 'withdrawUnbonded',
+    args: [slashingSpans],
     account: ensureHex(nominatorAddress),
   });
 
@@ -206,15 +126,22 @@ export const stopNomination = async (
   return txHash;
 };
 
-export const withdrawUnbondedTokens = async (
+export const updatePaymentDestination = async (
   nominatorAddress: string,
-  slashingSpans: number
+  paymentDestination: string
 ): Promise<AddressType> => {
+  const payee =
+    paymentDestination === PaymentDestination.Staked
+      ? PAYEE_STAKED
+      : paymentDestination === PaymentDestination.Stash
+      ? PAYEE_STASH
+      : PAYEE_CONTROLLER;
+
   const { request } = await evmPublicClient.simulateContract({
     address: StakingInterfacePrecompileAddress,
     abi: StakingInterfacePrecompileABI,
-    functionName: 'withdrawUnbonded',
-    args: [slashingSpans],
+    functionName: 'setPayee',
+    args: [payee],
     account: ensureHex(nominatorAddress),
   });
