+++ apps/tangle-dapp/utils/evm/index.ts
@@ -1,0 +1,3 @@
+export * from './bond';
+export * from './client';
+export * from './nominators';
